Refactor notes and feature settings flow
