+++ Voxt/Settings/FeatureSettingsComponents.swift
@@ -346,106 +346,6 @@ struct SettingsShortcutCaptureField: View {
     }
 }
 
-struct FeatureNoteShortcutRow: View {
-    let title: String
-    let detail: String
-    @Binding var shortcut: TranscriptionNoteTriggerSettings
-
-    @State private var isRecording = false
-    @State private var recorderMessage: String?
-    @State private var pendingCapturedHotkey: HotkeyPreference.Hotkey?
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(.subheadline.weight(.semibold))
-                Text(detail)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-
-            SettingsShortcutCaptureField(
-                title: LocalizedStringKey(title),
-                hotkey: pendingCapturedHotkey ?? shortcut.hotkey,
-                isRecording: isRecording,
-                isPendingConfirmation: pendingCapturedHotkey != nil,
-                distinguishModifierSides: false,
-                onFocus: {
-                    pendingCapturedHotkey = nil
-                    isRecording = true
-                },
-                onReset: {
-                    shortcut = .defaultShortcut
-                    pendingCapturedHotkey = nil
-                    isRecording = false
-                    recorderMessage = nil
-                },
-                onCancelPending: {
-                    pendingCapturedHotkey = nil
-                    isRecording = false
-                    recorderMessage = nil
-                },
-                onConfirmPending: {
-                    if let pendingCapturedHotkey {
-                        shortcut = TranscriptionNoteTriggerSettings(
-                            keyCode: pendingCapturedHotkey.keyCode,
-                            modifiers: pendingCapturedHotkey.modifiers,
-                            sidedModifiers: pendingCapturedHotkey.sidedModifiers
-                        )
-                    }
-                    self.pendingCapturedHotkey = nil
-                    isRecording = false
-                    recorderMessage = nil
-                }
-            )
-
-            if let recorderMessage, !recorderMessage.isEmpty {
-                Text(localized(recorderMessage))
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            } else if pendingCapturedHotkey != nil {
-                Text(localized("Shortcut captured. Press another shortcut to replace it, or choose Confirm / Cancel."))
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            } else if isRecording {
-                Text(localized("Type your shortcut now. Press Esc to cancel recording."))
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-
-            HotkeyRecorderView(
-                isRecording: $isRecording,
-                onCapture: { capturedHotkey in
-                    pendingCapturedHotkey = capturedHotkey
-                    recorderMessage = nil
-                },
-                onCancelCapture: {
-                    pendingCapturedHotkey = nil
-                    isRecording = false
-                    recorderMessage = nil
-                },
-                onRecorderMessageChange: { messageKey in
-                    DispatchQueue.main.async {
-                        recorderMessage = messageKey
-                    }
-                }
-            )
-            .frame(width: 0, height: 0)
-        }
-        .padding(14)
-        .background(
-            RoundedRectangle(cornerRadius: SettingsUIStyle.compactCornerRadius, style: .continuous)
-                .fill(SettingsUIStyle.groupedFillColor)
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: SettingsUIStyle.compactCornerRadius, style: .continuous)
-                .stroke(SettingsUIStyle.subtleBorderColor, lineWidth: 1)
-        )
-    }
-}
-
 struct FeaturePromptSection: View {
     let title: String
     @Binding var text: String
