+++ Voxt/Support/PasteboardSupport.swift
@@ -1,0 +1,7 @@
+import AppKit
+
+func copyStringToPasteboard(_ text: String) {
+    let pasteboard = NSPasteboard.general
+    pasteboard.clearContents()
+    pasteboard.setString(text, forType: .string)
+}
